Add TODOs, adds shell script for testing
Shell script takes a specified list (specify the list in the script itself) to run against all layout files it can find. This should help with bulk testing later, as we can simply loop this script if necessary and store the results somewhere to do whatever sort of training we want with.